+++ cmd/hidden_lake/service/pkg/settings/settings.go
@@ -59,8 +59,8 @@ const (
 )
 
 const (
-	CNetworkCapacity = (1 << 10) // 1024 hashes
-	CNetworkMaxConns = (1 << 8)  // 256 conns
+	CNetworkQueueSize = (1 << 10) // 1024 hashes
+	CNetworkMaxConns  = (1 << 8)  // 256 conns
 )
 
 const (