+++ cmd/hidden_lake/traffic/pkg/app/init_node.go
@@ -29,7 +29,7 @@ func initNode(pCfg config.IConfig, pWrapperDB database.IWrapperDB, pLogger logge
 		network.NewSettings(&network.SSettings{
 			FAddress:      pCfg.GetAddress().GetTCP(),
 			FMaxConnects:  hls_settings.CNetworkMaxConns,
-			FCapacity:     hls_settings.CNetworkCapacity,
+			FQueueSize:    hls_settings.CNetworkQueueSize,
 			FReadTimeout:  queueDuration,
 			FWriteTimeout: queueDuration,
 			FConnSettings: conn.NewSettings(&conn.SSettings{