+++ pkg/network/examples/broadcast/main.go
@@ -64,7 +64,7 @@ func handler(serviceName string) network.IHandlerF {
 func nodeSettings(serviceAddress string) network.ISettings {
 	return network.NewSettings(&network.SSettings{
 		FAddress:      serviceAddress,
-		FCapacity:     (1 << 10),
+		FQueueSize:    (1 << 10),
 		FMaxConnects:  2,
 		FConnSettings: connSettings(),
 		FWriteTimeout: time.Minute,