+++ pkg/network/examples/echo/main.go
@@ -68,7 +68,7 @@ func handler() network.IHandlerF {
 func nodeSettings(serviceAddress string) network.ISettings {
 	return network.NewSettings(&network.SSettings{
 		FAddress:      serviceAddress,
-		FCapacity:     (1 << 10),
+		FQueueSize:    (1 << 10),
 		FMaxConnects:  1,
 		FConnSettings: connSettings(),
 		FWriteTimeout: time.Minute,