+++ pkg/network/examples/ping-pong/main.go
@@ -78,7 +78,7 @@ func handler(serviceName string) network.IHandlerF {
 func nodeSettings(serviceAddress string) network.ISettings {
 	return network.NewSettings(&network.SSettings{
 		FAddress:      serviceAddress,
-		FCapacity:     (1 << 10),
+		FQueueSize:    (1 << 10),
 		FMaxConnects:  1,
 		FConnSettings: connSettings(),
 		FWriteTimeout: time.Minute,