+++ pkg/network/types.go
@@ -27,7 +27,7 @@ type INode interface {
 
 type ISettings interface {
 	GetAddress() string
-	GetCapacity() uint64
+	GetQueueSize() uint64
 	GetMaxConnects() uint64
 	GetReadTimeout() time.Duration
 	GetWriteTimeout() time.Duration